pc端修改

+++ website/src/main/java/com/gxa/jbgsw/website/controller/MyCollectionController.java
@@ -4,7 +4,9 @@ import com.gxa.jbgsw.business.protocol.dto.MyAttentionRequest;
 import com.gxa.jbgsw.business.protocol.dto.MyAttentionResponse;
 import com.gxa.jbgsw.business.protocol.dto.MyCollectionRequest;
 import com.gxa.jbgsw.business.protocol.dto.MyCollectionResponse;
+import com.gxa.jbgsw.common.exception.BizException;
 import com.gxa.jbgsw.common.utils.BaseController;
+import com.gxa.jbgsw.user.protocol.errcode.UserErrorCode;
 import com.gxa.jbgsw.website.feignapi.CollectionFeignApi;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -23,7 +25,12 @@ public class MyCollectionController extends BaseController {
     @ApiOperation("获取我的收藏")
     @PostMapping("/collection/pageQuery")
     MyCollectionResponse pageQuery(@RequestBody MyCollectionRequest myCollectionRequest){
-        myCollectionRequest.setCreateBy(this.getUserId());
+        Long userId = this.getUserId();
+        if(userId == null){
+            throw new BizException(UserErrorCode.LOGIN_SESSION_EXPIRE);
+        }
+
+        myCollectionRequest.setCreateBy(userId);
         MyCollectionResponse response = collectionFeignApi.queryMyCollections(myCollectionRequest);
 
         return response;
@@ -32,6 +39,11 @@ public class MyCollectionController extends BaseController {
     @ApiOperation(value = "取消收藏", notes = "取消收藏: 传收藏ID")
     @GetMapping("/collection/deleteById")
     public void deleteById(@RequestParam("id") Long id){
+        Long userId = this.getUserId();
+        if(userId == null){
+            throw new BizException(UserErrorCode.LOGIN_SESSION_EXPIRE);
+        }
+
         collectionFeignApi.deleteById(id);
     }
