+++ website/src/main/java/com/gxa/jbgsw/website/controller/MyPublishBillboardController.java
@@ -54,6 +54,11 @@ public class MyPublishBillboardController extends BaseController {
     @ApiOperation("获取我发布的榜单列表")
     @PostMapping("/user/center/queryMyPublish")
     MyPublishBillboardResponse queryMyPublish(@RequestBody MyPublishBillboardRequest request){
+        Long userId = this.getUserId();
+        if(userId == null){
+            throw new BizException(UserErrorCode.LOGIN_SESSION_EXPIRE);
+        }
+
         request.setUserId(this.getUserId());
         MyPublishBillboardResponse response = billboardFeignApi.queryMyPublish(request);
         List<MyPublishBillboardInfo> billboards = response.getBillboards();
@@ -76,6 +81,11 @@ public class MyPublishBillboardController extends BaseController {
     })
     @GetMapping("/user/center/deleteMyBillboard")
     public void deleteMyBillboard(@RequestParam("id") Long id){
+        Long userId = this.getUserId();
+        if(userId == null){
+            throw new BizException(UserErrorCode.LOGIN_SESSION_EXPIRE);
+        }
+
         Long[] ids = {id};
         billboardFeignApi.deleteBatchIds(ids);
     }
@@ -83,10 +93,18 @@ public class MyPublishBillboardController extends BaseController {
     @ApiOperation("我要发榜")
     @PostMapping("/user/center/addMyBillboard")
     void addMyBillboard(@RequestBody BillboardDTO billboardDTO) throws BizException {
-        billboardDTO.setCreateBy(this.getUserId());
+        Long userId = this.getUserId();
+        if(userId == null){
+            throw new BizException(UserErrorCode.LOGIN_SESSION_EXPIRE);
+        }
+
+        billboardDTO.setCreateBy(userId);
         billboardDTO.setCreateAt(new Date());
         // 设置默认的待揭榜
         billboardDTO.setStatus(0);
+        if(billboardDTO.getUnitName() == null){
+            billboardDTO.setUnitName(this.getUnitName());
+        }
 
         // 判断是否政府发布政府榜单
         // TODO: 2023/7/7 0007 现在的流程暂时不用判断
@@ -97,6 +115,11 @@ public class MyPublishBillboardController extends BaseController {
     @ApiOperation("编辑我发布的榜单信息")
     @PostMapping("/user/center/updateMyBillboard")
     void updateMyBillboard(@RequestBody BillboardDTO billboardDTO) throws BizException {
+        Long userId = this.getUserId();
+        if(userId == null){
+            throw new BizException(UserErrorCode.LOGIN_SESSION_EXPIRE);
+        }
+
         if(billboardDTO.getId() == null){
             throw new BizException(BusinessErrorCode.BUSINESS_PARAMS_ERROR);
         }
@@ -119,6 +142,7 @@ public class MyPublishBillboardController extends BaseController {
             });
         }
         detailInfo.setBillboardGains(billboardGainResponses);
+        detailInfo.setStatusName(BillboardStatusEnum.getNameByIndex(detailInfo.getStatus()));
 
         // 成果推荐: 根据揭榜单位
         List<BillboardHarvestRelatedResponse> havests = billboardHarvestRelatedFeignApi.getHarvestRecommend(id);
@@ -141,8 +165,13 @@ public class MyPublishBillboardController extends BaseController {
     @ApiOperation("我的技术经纪人评分")
     @PostMapping("/user/center/addAppraise")
     void addAppraise(@RequestBody TechEconomicManAppraiseDTO techEconomicManAppraiseDTO) throws BizException {
+        Long userId = this.getUserId();
+        if(userId == null){
+            throw new BizException(UserErrorCode.LOGIN_SESSION_EXPIRE);
+        }
+
         techEconomicManAppraiseDTO.setCreateAt(new Date());
-        techEconomicManAppraiseDTO.setCreateBy(this.getUserId());
+        techEconomicManAppraiseDTO.setCreateBy(userId);
         if(techEconomicManAppraiseDTO.isAnonymous()){
             // 匿名发表评价
             techEconomicManAppraiseDTO.setName(null);
